putting all publicly unused deserialization classes under IndixApiClientImpl class
