--- src/main/java/models/metadataResponse/BrandsResponse.java
@@ -1,12 +1,0 @@
-package models.metadataResponse;
-
-import models.ResponseBase;
-import models.metadataResponse.metadataResult.BrandsResult;
-
-public class BrandsResponse extends ResponseBase {
-    private BrandsResult result;
-
-    public BrandsResult getResult() {
-        return result;
-    }
-}
