--- src/main/java/models/productDetailsResponse/CatalogPremiumProductDetailsResponse.java
@@ -1,12 +1,0 @@
-package models.productDetailsResponse;
-
-import models.ResponseBase;
-import models.productDetailsResponse.productDetailsResult.CatalogPremiumProductDetailsResult;
-
-public class CatalogPremiumProductDetailsResponse extends ResponseBase {
-    private CatalogPremiumProductDetailsResult result;
-
-    public CatalogPremiumProductDetailsResult getResult() {
-        return result;
-    }
-}
