--- src/main/java/models/productDetailsResponse/CatalogStandardProductDetailsResponse.java
@@ -1,12 +1,0 @@
-package models.productDetailsResponse;
-
-import models.ResponseBase;
-import models.productDetailsResponse.productDetailsResult.CatalogStandardProductDetailsResult;
-
-public class CatalogStandardProductDetailsResponse extends ResponseBase {
-    private CatalogStandardProductDetailsResult result;
-
-    public CatalogStandardProductDetailsResult getResult() {
-        return result;
-    }
-}
