--- src/main/java/models/productDetailsResponse/SummaryProductDetailsResponse.java
@@ -1,12 +1,0 @@
-package models.productDetailsResponse;
-
-import models.ResponseBase;
-import models.productDetailsResponse.productDetailsResult.SummaryProductDetailsResult;
-
-public class SummaryProductDetailsResponse extends ResponseBase {
-    private SummaryProductDetailsResult result;
-
-    public SummaryProductDetailsResult getResult() {
-        return result;
-    }
-}
