--- src/main/java/models/productDetailsResponse/UniversalProductDetailsResponse.java
@@ -1,12 +1,0 @@
-package models.productDetailsResponse;
-
-import models.ResponseBase;
-import models.productDetailsResponse.productDetailsResult.UniversalProductDetailsResult;
-
-public class UniversalProductDetailsResponse extends ResponseBase {
-    private UniversalProductDetailsResult result;
-
-    public UniversalProductDetailsResult getResult() {
-        return result;
-    }
-}
