--- src/main/java/models/productHistoryResponse/ProductHistoryResponse.java
@@ -1,11 +1,0 @@
-package models.productHistoryResponse;
-
-import models.ResponseBase;
-
-public class ProductHistoryResponse extends ResponseBase {
-    private ProductHistoryResult result;
-
-    public ProductHistoryResult getResult() {
-        return result;
-    }
-}
