--- src/main/java/models/searchResponse/CatalogStandardSearchResponse.java
@@ -1,12 +1,0 @@
-package models.searchResponse;
-
-import models.ResponseBase;
-import models.searchResponse.searchResult.CatalogStandardSearchResult;
-
-public class CatalogStandardSearchResponse extends ResponseBase {
-    private CatalogStandardSearchResult result;
-
-    public CatalogStandardSearchResult getResult() {
-        return result;
-    }
-}
