--- src/main/java/models/searchResponse/OffersSearchResponse.java
@@ -1,12 +1,0 @@
-package models.searchResponse;
-
-import models.ResponseBase;
-import models.searchResponse.searchResult.OffersSearchResult;
-
-public class OffersSearchResponse extends ResponseBase {
-    private OffersSearchResult result;
-
-    public OffersSearchResult getResult() {
-        return result;
-    }
-}
