--- src/main/java/models/searchResponse/SummarySearchResponse.java
@@ -1,12 +1,0 @@
-package models.searchResponse;
-
-import models.ResponseBase;
-import models.searchResponse.searchResult.SummarySearchResult;
-
-public class SummarySearchResponse extends ResponseBase {
-    private SummarySearchResult result;
-
-    public SummarySearchResult getResult() {
-        return result;
-    }
-}
