--- src/main/java/models/searchResponse/UniversalSearchResponse.java
@@ -1,12 +1,0 @@
-package models.searchResponse;
-
-import models.ResponseBase;
-import models.searchResponse.searchResult.UniversalSearchResult;
-
-public class UniversalSearchResponse extends ResponseBase {
-    private UniversalSearchResult result;
-
-    public UniversalSearchResult getResult() {
-        return result;
-    }
-}
