--- src/main/java/models/suggestions/SuggestionsResponse.java
@@ -1,11 +1,0 @@
-package models.suggestions;
-
-import models.ResponseBase;
-
-public class SuggestionsResponse extends ResponseBase {
-    private SuggestionsResult result;
-
-    public SuggestionsResult getResult() {
-        return result;
-    }
-}
